# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ClinicSaathi API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface ErrorEnvelope {
  error: string;
}

export interface Doctor {
  id: string;
  email: string;
  name: string;
  clinicName: string;
  phone: string;
  address: string;
}

export interface DoctorUpdate {
  name: string;
  clinicName: string;
  phone: string;
  address: string;
}

export type PatientGender = (typeof PatientGender)[keyof typeof PatientGender];

export const PatientGender = {
  Male: "Male",
  Female: "Female",
  Other: "Other",
} as const;

export interface Patient {
  id: string;
  name: string;
  phone: string;
  age: number;
  gender: PatientGender;
  createdAt: string;
}

export type PatientWithStats = Patient & {
  totalVisits: number;
  lastVisitDate: string | null;
  hasActiveReminder: boolean;
};

export type NewPatientGender =
  (typeof NewPatientGender)[keyof typeof NewPatientGender];

export const NewPatientGender = {
  Male: "Male",
  Female: "Female",
  Other: "Other",
} as const;

export interface NewPatient {
  /** @minLength 1 */
  name: string;
  /**
   * @minLength 10
   * @maxLength 10
   */
  phone: string;
  /**
   * @minimum 0
   * @maximum 130
   */
  age: number;
  gender: NewPatientGender;
}

export interface Appointment {
  id: string;
  patientId: string;
  /** ISO date (YYYY-MM-DD) */
  appointmentDate: string;
  createdAt: string;
}

export interface MedicineReminder {
  id: string;
  patientId: string;
  /** ISO date (YYYY-MM-DD) */
  startDate: string;
  /** ISO date (YYYY-MM-DD) */
  endDate: string;
  durationDays: number;
  isActive: boolean;
  createdAt: string;
}

export interface PatientDetail {
  patient: Patient;
  totalVisits: number;
  lastVisitDate: string | null;
  nextAppointment: Appointment | null;
  activeReminder: MedicineReminder | null;
}

export interface NewAppointment {
  /** ISO date (YYYY-MM-DD) */
  appointmentDate: string;
}

export interface Prescription {
  id: string;
  patientId: string;
  photoUrl: string;
  createdAt: string;
}

export interface NewPrescription {
  photoUrl: string;
}

export interface NewReminder {
  /**
   * @minimum 1
   * @maximum 365
   */
  durationDays: number;
  isActive: boolean;
}

export interface TodayPatientRow {
  patientId: string;
  name: string;
  visitTime: string;
  nextAppointmentDate?: string | null;
  hasActiveReminder: boolean;
}

export interface DashboardSummary {
  todayPatients: number;
  todayAppointments: number;
  activeReminders: number;
  patientsThisMonth: number;
  todayPatientList: TodayPatientRow[];
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export type ListPatientsParams = {
  /**
   * Search by name or phone
   */
  q?: string;
};