# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ClinicSaathi API
 * OpenAPI spec version: 0.1.0
 */
import type { TodayPatientRow } from "./todayPatientRow";

export interface DashboardSummary {
  todayPatients: number;
  todayAppointments: number;
  activeReminders: number;
  patientsThisMonth: number;
  todayPatientList: TodayPatientRow[];
}