# lib/api-zod/src/generated/types/doctor.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ClinicSaathi API
 * OpenAPI spec version: 0.1.0
 */

export interface Doctor {
  id: string;
  email: string;
  name: string;
  clinicName: string;
  phone: string;
  address: string;
}