# lib/api-zod/src/generated/types/newAppointment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ClinicSaathi API
 * OpenAPI spec version: 0.1.0
 */

export interface NewAppointment {
  /** ISO date (YYYY-MM-DD) */
  appointmentDate: string;
}