# lib/api-zod/src/generated/types/newPatient.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ClinicSaathi API
 * OpenAPI spec version: 0.1.0
 */
import type { NewPatientGender } from "./newPatientGender";

export interface NewPatient {
  /** @minLength 1 */
  name: string;
  /**
   * @minLength 10
   * @maxLength 10
   */
  phone: string;
  /**
   * @minimum 0
   * @maximum 130
   */
  age: number;
  gender: NewPatientGender;
}