# lib/api-zod/src/generated/types/newReminder.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ClinicSaathi API
 * OpenAPI spec version: 0.1.0
 */

export interface NewReminder {
  /**
   * @minimum 1
   * @maximum 365
   */
  durationDays: number;
  isActive: boolean;
}