# lib/api-zod/src/generated/types/patient.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ClinicSaathi API
 * OpenAPI spec version: 0.1.0
 */
import type { PatientGender } from "./patientGender";

export interface Patient {
  id: string;
  name: string;
  phone: string;
  age: number;
  gender: PatientGender;
  createdAt: Date;
}