# lib/api-zod/src/generated/types/prescription.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ClinicSaathi API
 * OpenAPI spec version: 0.1.0
 */

export interface Prescription {
  id: string;
  patientId: string;
  photoUrl: string;
  createdAt: Date;
}