# lib/api-zod/src/generated/types/todayPatientRow.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ClinicSaathi API
 * OpenAPI spec version: 0.1.0
 */

export interface TodayPatientRow {
  patientId: string;
  name: string;
  visitTime: string;
  nextAppointmentDate?: string | null;
  hasActiveReminder: boolean;
}